+++ src/actions/queries.js
@@ -29,6 +29,13 @@ export function selectFacetValue(facetName, value, remove) {
 	return createNewQuery(Object.assign(part1, part2));
 }
 
+export function setFacetValues(facetValues) {
+	return createNewQuery({
+		type: "SET_FACET_VALUES",
+		facetValues: facetValues
+	});
+}
+
 export function newSearch() {
 	return createNewQuery({
 		type: "NEW_SEARCH"
@@ -39,6 +46,27 @@ export function changeSearchTerm(value) {
 	return createNewQuery({
 		type: "CHANGE_SEARCH_TERM",
 		value: value
+	});
+}
+
+export function changeFullTextSearchField(field, value) {
+	return createNewQuery({
+		type: "CHANGE_FULL_TEXT_SEARCH_TERM",
+		field: field,
+		value: value
+	});
+}
+
+export function setFullTextSearchFields(fullTextSearchParameters) {
+	return createNewQuery({
+		type: "SET_FULL_TEXT_SEARCH_TERMS",
+		fullTextSearchParameters: fullTextSearchParameters
+	});
+}
+
+export function removeFullTextSearchFields() {
+	return createNewQuery({
+		type: "REMOVE_FULL_TEXT_SEARCH_TERMS"
 	});
 }