+++ test/components/facets.js
@@ -9,7 +9,8 @@ describe("Facets", () => {
 		global.document = jsdom.jsdom('<!doctype html><html><body></body></html>');
 	});
 
-	it("should render properly", () => {
+	it("should render properly", function() {
+		this.timeout(4000);
 		// Can only import elements with insertCss after document is exposed as global
 		let Facets = require("../../src/components/facets");
 		let TextSearch = require("../../src/components/text-search");
@@ -47,8 +48,8 @@ describe("Facets", () => {
 		const { output } = setup();
 		expect(output.type).toBe("ul");
 		expect(output.props.className).toBe("hire-faceted-search-facets");
-		expect(output.props.children.length).toBe(3);
-		let [button, textSearch, childFacets] = output.props.children;
+		expect(output.props.children.length).toBe(5);
+		let [button, textSearch, ftSearchTop, childFacets, ftSearchBottom] = output.props.children;
 		expect(button.type).toBe("button");
 		expect(button.props.onClick()).toBe("new search clicked");
 		expect(textSearch.props.value).toBe("mock term");